backend/requirements.txt
fastapi
uvicorn
pytesseract
pillow
opencv-python
python-dotenv
httpx
cryptography
